refactor: rename search service file
